:recycle: Add Credential files
