+++ TransNotion/Service/Database/Local/LocalStore.swift
@@ -36,7 +36,3 @@ struct LocalStore<Coder: Codable & LocalStoreKey> {
         return decoded
     }
 }
-
-struct Credentials: Codable, LocalStoreKey {
-    var elements: [NotionOAuth.Credential]
-}
